Add bxt_hud_visible_landmarks
Generate a list of info_landmark entities, all of which has a PVS containing
the player. Then print out the target name associated with each of them for
identification.

It is not known if this is equivalent to "printing out landmarks that
are in the player's PVS" from the point of view of the engine, even though they
appear to be logically equivalent.