Set lower alpha and non-pulsating for SOLID_NOT triggers
Some triggers may be SOLID_NOT if inactive. The push and hurt triggers are the
most common examples. The c1a4i map provides examples of inactive push and hurt
triggers. This commit lowers the alpha and make them non-pulsating.

The trigger_transition is an exception: it is always SOLID_NOT, and so it makes
more sense for it to be displayed like a normal active trigger, as it never
changes into an active state.