Add TAS editor simulation via a second game
The main game creates a message queue and the second game connects to it
via bxt_tas_become_simulator_client. Then the main game constantly
writes to a temporary .hltas file and sends a console command to the
second game to run the .hltas. The second game sends back real player
data, which the main game then displays in the TAS editor. Access to the
temporary .hltas is synchronized with an interprocess mutex.