bxt_tas_new: Put first frame commands into load_command
When they are on the first frame, they cause the first frame to be\
simulated with wrong (old) settings. In practice this is not a problem
because the first frame is a loading frame anyway, however, this does
cause a resimulation in the TAS studio, and is a bit less clean.